+++ examples/wizard/main.rs
@@ -2,7 +2,12 @@
 *   creates a new webvh DID
 */
 
-use crate::{resolve::resolve, updating::edit_did, witness::witness_log_entry};
+use crate::{
+    did_web::{insert_also_known_as, save_did_web},
+    resolve::resolve,
+    updating::edit_did,
+    witness::witness_log_entry,
+};
 use affinidi_secrets_resolver::secrets::Secret;
 use affinidi_tdk::dids::{DID, KeyType};
 use ahash::HashMap;
@@ -22,6 +27,7 @@ use tracing::debug;
 use tracing_subscriber::filter;
 use url::Url;
 
+mod did_web;
 mod resolve;
 mod updating;
 mod witness;
@@ -38,6 +44,10 @@ struct ConfigInfo {
 
     /// Secrets relating to Witness Nodes
     pub witnesses: HashMap<String, Secret>,
+
+    /// Secrets relating to the DID Document itself
+    /// Key = ID
+    pub did_keys: HashMap<String, Secret>,
 }
 
 impl ConfigInfo {
@@ -184,6 +194,8 @@ async fn main() -> Result<()> {
 
 async fn create_new_did() -> Result<()> {
     let mut didwebvh = DIDWebVHState::default();
+    // Store keys that we want to use for updates
+    let mut authorization_secrets = ConfigInfo::default();
 
     // ************************************************************************
     // Step 1: Get the URLs for this DID
@@ -234,8 +246,8 @@ async fn create_new_did() -> Result<()> {
     // ************************************************************************
     // Step 3: Create the DID Document
     // ************************************************************************
-    let did_document = loop {
-        match create_did_document(&webvh_did) {
+    let mut did_document = loop {
+        match create_did_document(&webvh_did, &mut authorization_secrets) {
             Ok(doc) => break doc,
             Err(_) => {
                 println!(
@@ -266,8 +278,6 @@ async fn create_new_did() -> Result<()> {
     // ************************************************************************
     // Step 4: Configure Parameters
     // ************************************************************************
-    // Store keys that we want to use for updates
-    let mut authorization_secrets = ConfigInfo::default();
     authorizing_keys.iter().for_each(|key| {
         authorization_secrets.add_key(key);
     });
@@ -287,7 +297,23 @@ async fn create_new_did() -> Result<()> {
     debug!("Parameters: {parameters:#?}");
 
     // ************************************************************************
-    // Step 5: Create preliminary JSON Log Entry
+    // Step 6: Export this as a did:web?
+    // ************************************************************************
+
+    let export_did_web = if Confirm::with_theme(&ColorfulTheme::default())
+        .with_prompt("Would you like to export this DID as a did:web document as well?")
+        .default(true)
+        .interact()?
+    {
+        // Insert alsoKnownAs if not already there?
+        insert_also_known_as(&mut did_document, &webvh_did)?;
+        true
+    } else {
+        false
+    };
+
+    // ************************************************************************
+    // Step 7: Create preliminary JSON Log Entry
     // ************************************************************************
 
     let log_entry = didwebvh.create_log_entry(
@@ -304,7 +330,7 @@ async fn create_new_did() -> Result<()> {
     );
 
     // ************************************************************************
-    // Step 6: Validate the LogEntry
+    // Step 8: Validate the LogEntry
     // ************************************************************************
     // Validate the Log Entry
     let validated_params = log_entry.log_entry.verify_log_entry(None, None)?;
@@ -316,7 +342,7 @@ async fn create_new_did() -> Result<()> {
     );
 
     // ************************************************************************
-    // Step 7: Create the witness proofs if needed?
+    // Step 9: Create the witness proofs if needed?
     // ************************************************************************
     let mut witness_proofs = WitnessProofCollection::default();
     let new_proofs = witness_log_entry(
@@ -366,6 +392,11 @@ async fn create_new_did() -> Result<()> {
                 style("Witness Proofs saved to :").color256(69),
                 style([start, "-witness.json"].concat()).color256(214),
             );
+        }
+
+        // Export did:web document?
+        if export_did_web {
+            save_did_web(log_entry)?;
         }
     }
 
@@ -608,7 +639,7 @@ pub fn get_keys() -> Result<Vec<Secret>> {
 }
 
 // Create DID Document
-fn create_did_document(webvh_did: &str) -> Result<Value> {
+fn create_did_document(webvh_did: &str, config_info: &mut ConfigInfo) -> Result<Value> {
     println!(
         "{} {}",
         style("Create a DID Document for:").color256(69),
@@ -641,7 +672,7 @@ fn create_did_document(webvh_did: &str) -> Result<Value> {
     }
 
     // Add Verification Methods
-    get_verification_methods(webvh_did, &mut did_document);
+    get_verification_methods(webvh_did, &mut did_document, config_info);
 
     println!();
     println!(
@@ -737,7 +768,7 @@ fn also_known_as() -> Vec<String> {
 }
 
 // Create Verification Methods
-fn get_verification_methods(webvh_did: &str, doc: &mut Value) {
+fn get_verification_methods(webvh_did: &str, doc: &mut Value, config_info: &mut ConfigInfo) {
     let mut key_id: u32 = 0;
     let mut success_count: u32 = 0;
 
@@ -820,6 +851,8 @@ fn get_verification_methods(webvh_did: &str, doc: &mut Value) {
                     _ => {}
                 }
             }
+            // Add to secrets
+            config_info.did_keys.insert(vm_id, secret);
         }
         if success_count > 0
             && !Confirm::with_theme(&ColorfulTheme::default())
@@ -836,6 +869,7 @@ fn get_verification_methods(webvh_did: &str, doc: &mut Value) {
 fn create_key(id: &str) -> Secret {
     let items = vec![
         KeyType::Ed25519.to_string(),
+        "X25519".to_string(),
         KeyType::P256.to_string(),
         KeyType::Secp256k1.to_string(),
         KeyType::P384.to_string(),
@@ -848,8 +882,14 @@ fn create_key(id: &str) -> Secret {
         .interact()
         .unwrap();
 
-    let (_, mut secret) =
-        DID::generate_did_key(KeyType::try_from(items[selection].as_str()).unwrap()).unwrap();
+    let mut secret = if selection == 1 {
+        let (_, secret) = DID::generate_did_key(KeyType::Ed25519).unwrap();
+        secret.to_x25519().unwrap()
+    } else {
+        DID::generate_did_key(KeyType::try_from(items[selection].as_str()).unwrap())
+            .unwrap()
+            .1
+    };
 
     secret.id = id.to_string();
     secret