+++ examples/wizard/updating/mod.rs
@@ -2,7 +2,9 @@
 *   Tasks relating to editing an existing webvh DID go here
 */
 use crate::{
-    ConfigInfo, edit_did_document,
+    ConfigInfo,
+    did_web::save_did_web,
+    edit_did_document,
     updating::{
         authorization::update_authorization_keys, portable::migrate_did, revoke::revoke_did,
         watchers::modify_watcher_params, witness::modify_witness_params,
@@ -139,36 +141,55 @@ pub async fn edit_did() -> Result<()> {
                         .witness_proofs
                         .save_to_file(&[file_name_prefix, "-witness.json"].concat())?;
                 }
+
+                if Confirm::with_theme(&ColorfulTheme::default())
+                    .with_prompt("Export latest state to a did:web document?")
+                    .default(true)
+                    .interact()
+                    .unwrap()
+                {
+                    save_did_web(new_entry)?;
+                }
+
                 break;
             }
             1 => {
                 // DID Portability
-                migrate_did(&mut webvh_state, &mut config_info)?;
+                if migrate_did(&mut webvh_state, &mut config_info)? {
+                    let new_entry = webvh_state.log_entries.last().ok_or_else(|| {
+                        DIDWebVHError::LogEntryError("No new LogEntry created".to_string())
+                    })?;
 
-                let new_entry = webvh_state.log_entries.last().ok_or_else(|| {
-                    DIDWebVHError::LogEntryError("No new LogEntry created".to_string())
-                })?;
+                    let new_proofs = witness_log_entry(
+                        &mut webvh_state.witness_proofs,
+                        new_entry,
+                        &new_entry.get_active_witnesses(),
+                        &config_info,
+                    )?;
 
-                let new_proofs = witness_log_entry(
-                    &mut webvh_state.witness_proofs,
-                    new_entry,
-                    &new_entry.get_active_witnesses(),
-                    &config_info,
-                )?;
+                    // Save info to files
+                    new_entry.log_entry.save_to_file(&file_path)?;
+                    config_info.save_to_file(&[file_name_prefix, "-secrets.json"].concat())?;
+                    println!(
+                        "{}",
+                        style("Successfully created new LogEntry")
+                            .color256(34)
+                            .blink()
+                    );
+                    if new_proofs.is_some() {
+                        webvh_state
+                            .witness_proofs
+                            .save_to_file(&[file_name_prefix, "-witness.json"].concat())?;
+                    }
 
-                // Save info to files
-                new_entry.log_entry.save_to_file(&file_path)?;
-                config_info.save_to_file(&[file_name_prefix, "-secrets.json"].concat())?;
-                println!(
-                    "{}",
-                    style("Successfully created new LogEntry")
-                        .color256(34)
-                        .blink()
-                );
-                if new_proofs.is_some() {
-                    webvh_state
-                        .witness_proofs
-                        .save_to_file(&[file_name_prefix, "-witness.json"].concat())?;
+                    if Confirm::with_theme(&ColorfulTheme::default())
+                        .with_prompt("Export latest state to a did:web document?")
+                        .default(true)
+                        .interact()
+                        .unwrap()
+                    {
+                        save_did_web(new_entry)?;
+                    }
                 }
 
                 break;