+++ examples/wizard/updating/portable.rs
@@ -10,13 +10,11 @@ use anyhow::{Result, anyhow, bail};
 use console::style;
 use dialoguer::{Confirm, Input, theme::ColorfulTheme};
 use didwebvh_rs::{DIDWebVHState, parameters::Parameters, url::WebVHURL};
-use iref::IriBuf;
-use ssi::dids::Document;
-use std::str::FromStr;
+use serde_json::Value;
 use url::Url;
 
 /// Revokes a webvh DID method
-pub fn migrate_did(didwebvh: &mut DIDWebVHState, secrets: &mut ConfigInfo) -> Result<()> {
+pub fn migrate_did(didwebvh: &mut DIDWebVHState, secrets: &mut ConfigInfo) -> Result<bool> {
     let Some(log_entry) = didwebvh.log_entries.last() else {
         bail!("There must at least be a first LogEntry for this DID to migrate it");
     };
@@ -29,9 +27,11 @@ pub fn migrate_did(didwebvh: &mut DIDWebVHState, secrets: &mut ConfigInfo) -> Re
         .get_state()
         .get("id")
         .ok_or_else(|| anyhow::anyhow!("DID not found in the log entry state"))?
-        .as_str();
+        .as_str()
+        .ok_or_else(|| anyhow::anyhow!("DID is not a string"))?
+        .to_string();
 
-    let did_url = WebVHURL::parse_did_url(did.unwrap())?;
+    let did_url = WebVHURL::parse_did_url(&did)?;
 
     println!(
         "\n{}",
@@ -68,18 +68,23 @@ pub fn migrate_did(didwebvh: &mut DIDWebVHState, secrets: &mut ConfigInfo) -> Re
         .default(true)
         .interact()?
     {
-        return Ok(());
+        return Ok(false);
     }
 
     // Modify the DID Doc and create new LogEntry
     let did_doc: String = serde_json::to_string(&log_entry.get_state())?;
     let new_did_doc = did_doc.replace(&did_url.to_string(), &new_did_url.to_string());
-    let mut new_did_doc: Document = serde_json::from_str(&new_did_doc)?;
+    let mut new_did_doc: Value = serde_json::from_str(&new_did_doc)?;
 
     // Add to alsoKnownAs
-    new_did_doc
-        .also_known_as
-        .push(IriBuf::from_str(did.unwrap())?);
+    if let Some(alias) = new_did_doc.get_mut("alsoKnownAs") {
+        alias.as_array_mut().unwrap().push(Value::String(did));
+    } else {
+        new_did_doc.as_object_mut().unwrap().insert(
+            "alsoKnownAs".to_string(),
+            Value::Array(vec![Value::String(did)]),
+        );
+    }
 
     println!(
         "{}",
@@ -92,7 +97,7 @@ pub fn migrate_did(didwebvh: &mut DIDWebVHState, secrets: &mut ConfigInfo) -> Re
         .interact()?
     {
         println!("{}", style("Migration aborted!").color256(141));
-        return Ok(());
+        return Ok(false);
     }
 
     // Create new LogEntry for this migration
@@ -110,14 +115,8 @@ pub fn migrate_did(didwebvh: &mut DIDWebVHState, secrets: &mut ConfigInfo) -> Re
     };
 
     didwebvh
-        .create_log_entry(
-            None,
-            &serde_json::to_value(new_did_doc)
-                .map_err(|e| anyhow!("Couldn't convert DID Doc to JSON: {}", e))?,
-            &new_params,
-            signing_key,
-        )
+        .create_log_entry(None, &new_did_doc, &new_params, signing_key)
         .map_err(|e| anyhow!("Couldn't create LogEntry: {}", e))?;
 
-    Ok(())
+    Ok(true)
 }